Incorrect MD5
With help from the specification I changed the padding functionality, which solved some of my problems but now it's still not outputting the correct hash.
I'm thinking this could maybe be an endian issue, but I'm still uncertain.

A lot of tidying and more references to be added.

I'm also unsure whether I should revert to the previous padding state but ultimately feel this is the best approach, epecially as most implementations that I've seen online apprach the MD as done above