Revert Again..
The code I had initially was generating new hash values each time, so I've reverted back to the code from earlier.
It still has the issue described in a commit from earlier on, but at least it's consistently wrong.
Added references to where I've adapted the code from.
Also regenerated and included the headers.
Still have a few days to figure it out so hoping for a breakthrough